Move UserReports view and url to users app.
