+++ taskr/tasks/views.py
@@ -1,5 +1,4 @@
 from django.contrib.auth import get_user_model
-from django.db.models import Q
 from django.shortcuts import get_object_or_404
 
 from rest_framework import generics, status
@@ -38,6 +37,8 @@ class TaskListCreate(generics.GenericAPIView):
     '''
     permission_classes = (IsAuthenticated,)
     pagination_class = CustomPagination
+    # serializer_class = TaskSerializer
+    # queryset = Task.objects.all()
 
     def get(self, request, format=None):
         '''
@@ -242,47 +243,3 @@ class TaskEventLogList(APIView):
         logs_serializer = TaskEventLogSerializer(logs, many=True)
 
         return Response(logs_serializer.data)
-
-
-class UserReports(APIView):
-    '''
-    Reporting task info for a user.
-
-    Counts tasks for each user that are
-      - created
-      - assigned
-      - completed
-      - incompleted
-
-    * Requires token authentication.
-    '''
-    permission_classes = (IsAuthenticated,)
-
-    def get(self, request, username):
-        user = get_object_or_404(User, username=username)
-
-        # Get queryset of tasks that user has created or assigned to.
-        tasks = Task.objects.all()
-
-        # Count the created tasks for a user.
-        created_count = tasks.filter(reporter=user).count()
-
-        # Count the assigned, completed, incompleted tasks for a user.
-        # Derived from queryset of assigned tasks.
-        assigned_tasks = tasks.filter(assignee=user)
-        assigned_count = assigned_tasks.count()
-        completed_count = assigned_tasks.filter(
-            status=enums.STATUS_DONE
-        ).count()
-        incompleted_count = assigned_tasks.filter(
-            status__in=[enums.STATUS_TODO, enums.STATUS_IN_PROGRESS]
-        ).count()
-
-        # Create response object.
-        response = {}
-        response['created'] = created_count
-        response['assigned'] = assigned_count
-        response['completed'] = completed_count
-        response['incompleted'] = incompleted_count
-
-        return Response(response, status=status.HTTP_200_OK)
